try to fix unlit cylinder caps
