+++ rt/multi_light.rt
@@ -4,5 +4,6 @@ L	-30,30,0		0.9		255,0,0
 L	0,60,0			0.9		0,255,0
 L	30,30,0		0.9		0,0,255
 sp	0,10,20				25				200,200,200
+cy	-30,10,10	0,1,0	10 30	200,200,200
 pl	0,-10,0		0,1,0			200,200,200
